+++ src/app/outbound/auth_ctx/handlers/sign_up.py
@@ -84,6 +84,8 @@ class SignUp:
             now=now,
         )
         self._user_tx_storage.add(user)
+        events = user.collect_events()
+        await self._event_dispatcher.stage(events)
         try:
             await self._flusher.flush()
         except UsernameAlreadyExistsError:
@@ -94,7 +96,7 @@ class SignUp:
             raise
 
         await self._transaction_manager.commit()
-        await self._event_dispatcher.dispatch(user.collect_events())
+        await self._event_dispatcher.dispatch(events)
 
         logger.info("Sign up: done.")